KEEP OR DEPRECATE THE FETCH EXTENSIONS
Not a lot of impact either way
But it would keep me from going to the next level
I can fix these in the future if I use them.
FWIW the subroutines in question are not nor have they ever been used
in production ~BUT they may be useful in the OLDER Cocoa-Bindings where
they came from (i.e. In a multi column view where the filter was on the
topic of a given column)